Making the form have a better layout
